requirements.txt
discord.py
aiosqlite
python-dotenv
openai
Pillow
imagehash
torch
torchvision
aiohttp
